requirements.txt
PyQt6
google-genai
yt-dlp
Pillow
